begin restructuring
