+++ backend/.dockerignore
@@ -1,0 +1,3 @@
+__pycache__/
+.mypy_cache/
+venv/
