+++ backend/scoreboard/__init__.py
@@ -1,1 +1,1 @@
-from .main import app
+from .app import app
