+++ backend/scoreboard/app.py
@@ -1,78 +1,16 @@
 import asyncio
 import json
-import os
 from typing import Any, Dict, List, Optional
 
-import aredis
 from fastapi import FastAPI, HTTPException, status
 from fastapi.responses import Response
-from pydantic import BaseModel
-from watchgod import Change, awatch
+
+from .common import current_round, redis
+from .models import TeamState
 
 app = FastAPI()
-REDIS_HOST = os.getenv("REDIS_HOST", "127.0.0.1")
-DATA_DIR = os.getenv("DATA_DIR", "../data")
-redis: aredis.StrictRedis = aredis.StrictRedis(host=REDIS_HOST)
 UPDATE_EVENT: asyncio.Event = asyncio.Event()
 scoreboard_cache: Dict[int, bytes] = dict()
-
-
-class ServiceDetail(BaseModel):
-    ServiceId: int
-    AttackPoints: float
-    LostDefensePoints: float
-    ServiceLevelAgreementPoints: float
-    ServiceStatus: str
-    Message: str
-
-
-class ScoreboardTeam(BaseModel):
-    Name: str
-    TeamId: int
-    TotalPoints: float
-    AttackPoints: float
-    LostDefensePoints: float
-    ServiceLevelAgreementPoints: float
-    ServiceDetails: List[ServiceDetail]
-
-
-class FirstBlood(BaseModel):
-    TeamId: int
-    Timestamp: str
-    RoundId: int
-    StoreDescription: Optional[str]
-    StoreIndex: int
-
-
-class ScoreboardService(BaseModel):
-    ServiceId: int
-    ServiceName: str
-    MaxStores: int
-    FirstBloods: List[FirstBlood]
-
-
-class JsonScoreboard(BaseModel):
-    CurrentRound: Optional[int]
-    StartTimeEpoch: Optional[int]
-    EndTimeEpoch: Optional[int]
-    Services: List[ScoreboardService]
-    Teams: List[ScoreboardTeam]
-
-
-class TeamState(BaseModel):
-    Round: int
-    TotalPoints: float
-    AttackPoints: float
-    LostDefensePoints: float
-    ServiceLevelAgreementPoints: float
-    ServiceDetails: List[ServiceDetail]
-
-
-async def current_round() -> Optional[int]:
-    r = await redis.get("max_round")
-    if not r:
-        return None
-    return int(r.decode())
 
 
 async def get_scoreboard(round: Optional[int] = None) -> Optional[bytes]:
@@ -188,53 +126,7 @@ async def get_team(team_id: int) -> Response:
 
 @app.on_event("startup")
 async def startup() -> None:
-    base_path = os.path.join(DATA_DIR, "scoreboard.json")
-    await parse_scoreboard(base_path)
-    r = await current_round()
-    if r is not None:
-        for i in range(0, r + 1):
-            sb_path = os.path.join(DATA_DIR, f"scoreboard{i}.json")
-            await parse_scoreboard(sb_path)
-
-    asyncio.create_task(create_watch())
     asyncio.create_task(handle_updates())
-
-
-async def create_watch() -> None:
-    async for changes in awatch(DATA_DIR):
-        for c in changes:
-            if c[0] == Change.added or c[0] == Change.modified:
-                await parse_scoreboard(c[1])
-
-
-async def parse_scoreboard(file_: str) -> None:
-    basename = os.path.basename(file_)
-    if not basename.startswith("scoreboard") or not basename.endswith(".json"):
-        return
-    try:
-        obj = json.load(open(file_, "r"))
-        sb = JsonScoreboard(**obj)
-        print(f"Loading scoreboard for round {sb.CurrentRound}")
-
-        if not sb.CurrentRound:
-            return
-
-        entry = await redis.get(f"sb_{sb.CurrentRound}")
-        if not entry:
-            await redis.set(f"sb_{sb.CurrentRound}", sb.json())
-
-        entry = await redis.get("max_round")
-        print(f"previous: max_round: {entry}")
-        if not entry or int(entry.decode()) < sb.CurrentRound:
-            await redis.set("max_round", sb.CurrentRound)
-            await redis.publish("notifications", sb.CurrentRound)
-
-        for t in sb.Teams:
-            await redis.set(f"team_exists_${t.TeamId}", True)
-    except FileNotFoundError as e:
-        print(f"Failed to load scoreboard: {file_}, {str(e)}")
-    except json.JSONDecodeError as e:
-        print(f"Failed to parse scoreboard: {file_}, {str(e)}")
 
 
 async def handle_updates() -> None:
